+++ database/migrations/2017_11_16_172658_add_class_name_to_imports_table.php
@@ -1,0 +1,31 @@
+<?php
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class AddClassNameToImportsTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('imports', function (Blueprint $table) {
+            $table->string('class_name');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('imports', function (Blueprint $table) {
+            $table->dropColumn('class_name');
+        });
+    }
+}