+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file[^1].
 ### Added
 - Importer module
 
+### Fixed
+- Database seeders
+
 ## [1.2.8] - 2017-10-24
 ### Added
 - CONTRIBUTING.md