+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file[^1].
 - Importer module
 - Czech language translations
 
+## [1.2.9] - 2017-12-05
+### Changed
+- Enabled reproduction orders with mounting / framing
+
 ## [1.2.8] - 2017-10-24
 ### Added
 - CONTRIBUTING.md